requirements.txt
pyautogui
python3-xlib
Pillow
mouse
keyboard